# src/main/java/com/redhat/lightwell/model/dto/UpdateCustomerRequest.java
package com.redhat.lightwell.model.dto;

import javax.validation.constraints.Email;

public class UpdateCustomerRequest {

    private String firstName;

    private String lastName;

    @Email(message = "Email must be valid")
    private String email;

    private String phoneNumber;

    public UpdateCustomerRequest() {
    }

    public String getFirstName() {
        return firstName;
    }

    public void setFirstName(String firstName) {
        this.firstName = firstName;
    }

    public String getLastName() {
        return lastName;
    }

    public void setLastName(String lastName) {
        this.lastName = lastName;
    }

    public String getEmail() {
        return email;
    }

    public void setEmail(String email) {
        this.email = email;
    }

    public String getPhoneNumber() {
        return phoneNumber;
    }

    public void setPhoneNumber(String phoneNumber) {
        this.phoneNumber = phoneNumber;
    }
}